[MOD]- modify with a servic pattan
